Fix package refactors
